+++ scanner/__tests__/fixtures/r002-fail.vue
@@ -1,0 +1,5 @@
+<template>
+  <el-table :data="list">
+    <el-table-column prop="name" label="名称" align="center" />
+  </el-table>
+</template>