+++ scanner/__tests__/fixtures/r016-fail.vue
@@ -1,0 +1,9 @@
+<template>
+  <div class="box">hello</div>
+</template>
+<style scoped>
+.box {
+  color: #409eff;
+  background-color: #f56c6c;
+}
+</style>